Implemented `embed` method on PostBuilder API and refactored embeds
- EmbedInterface was introduced. All embeds were refactored to implement EmbedInterface.
This interface extends two essential interfaces - JsonSerializable and Stringable,
which are mandatory for Lexicon implementations. The reason is that when building a post with the PostBuilder API,
each entity that can be used as an embed is a lexicon, and a lexicon should be serializable to an object or array type
according to SDK design policy when necessary. The Lexicon interface to be introduced
in the future will also extend these two interfaces, and all necessary entities will be refactored.

- An embed method was added to the PostBuilder API, related tests were refactored/written.